hostgroup: Fix parent hostgroup handling
Foreman divides hostgroups from parent hostgroups by a '/' in the title,
e.g.

  foo/green
  webtier/green
  dbtier/green
  secure/webtier/green

while the name of the hostgroup would always be "green".

So in order to not pick a hostgroup at random search by title and if the
hostgroup name contains slashes wire up the correct parent host group so
we inherit it's parameters.