Allow to update more hostgroup parameters
The current code excluded quiet some parameters. Updating works (at
least with Foreman 1.10 onward) if we convert the ids to strings.

We then need to filter out all _name attributes of the returned
hostgroup (we do this via a white list) and convert the ids to strings
for comparison as well to not report items as changed which arent.